move component SearchBar in App.jsx
